Update program 07.py

+++ ProgramBatch2/program 07.py
@@ -1,2 +1,2 @@
-numbers = [int(input(f"Enter number {i+1}: ")) for i in range(10)]
+numbers = [float(input(f"Enter number {i+1}: ")) for i in range(10)]
 print("Count of even numbers:", sum(1 for n in numbers if n % 2 == 0))
